@@ -29,59 +29,62 @@
     "sonar56x": true
   },
   "dependencies": {
-    "@angular/animations": "^18.2.12",
-    "@angular/cdk": "^18.2.14",
-    "@angular/common": "^18.2.12",
-    "@angular/compiler": "^18.2.12",
-    "@angular/core": "^18.2.12",
-    "@angular/forms": "^18.2.12",
-    "@angular/localize": "^18.2.12",
-    "@angular/material": "^18.2.14",
-    "@angular/platform-browser": "^18.2.12",
-    "@angular/platform-browser-dynamic": "^18.2.12",
-    "@angular/router": "^18.2.12",
-    "@babel/core": "^7.18.10",
-    "@ngx-translate/core": "^15.0.0",
+    "@angular/animations": "^20.3.9",
+    "@angular/cdk": "^20.2.10",
+    "@angular/common": "^20.3.9",
+    "@angular/compiler": "^20.3.9",
+    "@angular/core": "^20.3.9",
+    "@angular/forms": "^20.3.9",
+    "@angular/localize": "^20.3.9",
+    "@angular/material": "^20.2.10",
+    "@angular/platform-browser": "^20.3.9",
+    "@angular/platform-browser-dynamic": "^20.3.9",
+    "@angular/router": "^20.3.9",
+    "@babel/core": "^7.28.5",
+    "@ngx-translate/core": "^17.0.0",
     "@ngx-translate/http-loader": "^8.0.0",
-    "@oblique/oblique": "^12.0.2",
-    "@popperjs/core": "^2.11.6",
-    "@quadrel-services/qd-auth": "^16.1.8",
-    "@types/istanbul-lib-report": "^3.0.0",
-    "@types/istanbul-reports": "^3.0.1",
-    "@types/ws": "^8.5.4",
+    "@oblique/oblique": "^14.0.2",
+    "@popperjs/core": "^2.11.8",
+    "@quadrel-services/qd-auth": "^19.0.12",
+    "@types/istanbul-lib-report": "^3.0.3",
+    "@types/istanbul-reports": "^3.0.4",
+    "@types/ws": "^8.18.1",
     "ajv": "^8.12.0",
-    "angular-oauth2-oidc": "^17.0.2",
+    "angular-oauth2-oidc": "^20.0.2",
     "date-fns": "^3.6.0",
     "fast-deep-equal": "^3.1.3",
+    "js-cookie": "^3.0.5",
     "json-schema-traverse": "^1.0.0",
     "jwt-decode": "^4.0.0",
     "rxjs": "^7.8.0",
-    "sass": "^1.58.3",
-    "tslib": "^2.5.0",
-    "zone.js": "~0.14.10"
+    "sass": "^1.93.2",
+    "tslib": "^2.8.1",
+    "update": "^0.7.4",
+    "zone.js": "^0.15.1"
   },
   "devDependencies": {
-    "@angular-builders/jest": "^17.0.3",
-    "@angular-devkit/build-angular": "^18.2.12",
-    "@angular/cli": "^18.2.12",
-    "@angular/compiler-cli": "^18.2.12",
+    "@angular-builders/jest": "^20.0.0",
+    "@angular/build": "^20.3.8",
+    "@angular/cli": "^20.3.8",
+    "@angular/compiler-cli": "^20.3.9",
     "@bbc/apache2-license-checker": "^1.1.4",
     "@ngtools/json-schema": "^1.1.0",
-    "@types/jest": "^29.5.12",
-    "codelyzer": "^6.0.0",
+    "@schematics/angular": "^20.3.7",
+    "@types/jest": "^29.5.14",
+    "codelyzer": "^6.0.2",
     "commander": "^13.1.0",
     "conventional-changelog": "^6.0.0",
     "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.5.0",
-    "jest-preset-angular": "^12.2.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "jest-preset-angular": "^12.2.6",
     "jest-sonar-reporter": "^2.0.0",
     "license-checker": "^25.0.1",
+    "node-gyp": "^10.3.1",
     "protractor": "~7.0.0",
     "simple-git": "3.24.0",
     "ts-jest": "29.1.3",
     "ts-node": "^10.9.2",
     "tslint": "~6.1.0",
-    "typescript": "5.4.5",
-    "node-gyp": "^10.1.0"
+    "typescript": "5.9.3"
   }
 }
