@@ -2,7 +2,8 @@ import {Component} from '@angular/core';
 
 @Component({
 	selector: 'app-error-details-page',
-	templateUrl: './error-details-page.component.html'
+	templateUrl: './error-details-page.component.html',
+	standalone: false
 })
 export class ErrorDetailsPageComponent {
 }
