@@ -1,8 +1,9 @@
-import { Component } from '@angular/core';
+import {Component} from '@angular/core';
 
 @Component({
-  selector: 'app-error-group-page',
-  templateUrl: './error-group-page.component.html',
+	selector: 'app-error-group-page',
+	templateUrl: './error-group-page.component.html',
+	standalone: false
 })
 export class ErrorGroupPageComponent {
 }
