@@ -2,7 +2,8 @@ import {Component} from '@angular/core';
 
 @Component({
 	selector: 'app-error-list-page',
-	templateUrl: './error-list-page.component.html'
+	templateUrl: './error-list-page.component.html',
+	standalone: false
 })
 export class ErrorListPageComponent {
 }
