@@ -1,15 +1,17 @@
-import { Component } from '@angular/core';
-import { HttpStatusCode } from '@angular/common/http';
+import {Component} from '@angular/core';
+import {HttpStatusCode} from '@angular/common/http';
 import {TranslateService} from '@ngx-translate/core';
 
 @Component({
-  selector: 'app-forbidden-page',
-  styleUrls: ['./forbidden-page.component.scss'],
-  templateUrl: './forbidden-page.component.html'
+	selector: 'app-forbidden-page',
+	styleUrls: ['./forbidden-page.component.scss'],
+	templateUrl: './forbidden-page.component.html',
+	standalone: false
 })
 export class ForbiddenPageComponent {
-  httpStatusCode = HttpStatusCode.Forbidden;
+	httpStatusCode = HttpStatusCode.Forbidden;
 
-	constructor(readonly translate: TranslateService) {}
+	constructor(readonly translate: TranslateService) {
+	}
 
 }
