@@ -4,7 +4,8 @@ import {MAT_DIALOG_DATA, MatDialogRef} from '@angular/material/dialog';
 @Component({
 	selector: 'app-closing-reason-dialog',
 	templateUrl: './closing-reason-dialog.component.html',
-	styleUrls: ['./closing-reason-dialog.component.scss']
+	styleUrls: ['./closing-reason-dialog.component.scss'],
+	standalone: false
 })
 export class ClosingReasonDialogComponent {
 
