@@ -4,7 +4,8 @@ import {MAT_DIALOG_DATA, MatDialogRef} from '@angular/material/dialog';
 @Component({
 	selector: 'app-confirmation-dialog',
 	templateUrl: './confirmation-dialog.component.html',
-	styleUrls: ['./confirmation-dialog.component.scss']
+	styleUrls: ['./confirmation-dialog.component.scss'],
+	standalone: false
 })
 export class ConfirmationDialogComponent {
 
